requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
sentence-transformers
pypdf
langchain-text-splitters
langchain
langchain_classic
langchain-openai
langchain-community
pgvector
python-dotenv
streamlit
requests
uuid